applied plugins to :feature, :app
