+++ build-logic/convention/src/main/kotlin/com/f776/convention/configureAndroidApplication.kt
@@ -4,13 +4,15 @@ import com.android.build.api.dsl.ApplicationExtension
 import org.gradle.api.Project
 
 internal fun Project.configureAndroidApplication(
-    extension: ApplicationExtension
+    extension: ApplicationExtension,
+    versionCodeProperty: Int,
+    versionNameProperty: String
 ) = extension.apply {
     defaultConfig {
-        applicationId = "com.f776"
+        applicationId = "com.kotlearn.minesweeperk"
         targetSdk = libs.findVersion("android-targetSdk").get().requiredVersion.toInt()
-        versionCode = 1
-        versionName = "1.0"
+        versionCode = versionCodeProperty
+        versionName = versionNameProperty
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
