+++ build-logic/convention/src/main/kotlin/com/f776/convention/configureCompose.kt
@@ -20,6 +20,7 @@ internal fun Project.configureCompose(
             implementation(composeDeps.components.resources)
             implementation(composeDeps.components.uiToolingPreview)
             implementation(composeDeps.material3)
+            implementation(composeDeps.materialIconsExtended)
             implementation(composeDeps.uiTest)
         }
