+++ composeApp/build.gradle.kts
@@ -1,9 +1,3 @@
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
-import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import java.io.FileInputStream
-import java.util.Properties
-
 plugins {
     alias(libs.plugins.kotlearn.kotlinMultiplatform)
     alias(libs.plugins.kotlearn.androidApplication)
@@ -11,38 +5,7 @@ plugins {
     alias(libs.plugins.kotlearn.desktopApplication)
 }
 
-val versionPropertiesInputStream = FileInputStream("$rootDir/versions.properties")
-val versionProperties = Properties().apply {
-    load(versionPropertiesInputStream)
-}
-val versionCodeProperty = versionProperties.getProperty("versionCode").toInt()
-val versionMajorProperty = versionProperties.getProperty("versionMajor").toInt()
-val versionMinorProperty = versionProperties.getProperty("versionMinor").toInt()
-val versionPatchProperty = versionProperties.getProperty("versionPatch").toInt()
-
-val versionNameProperty = "$versionMajorProperty.$versionMinorProperty.$versionPatchProperty"
-
 kotlin {
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
-        compilerOptions {
-            jvmTarget.set(JvmTarget.JVM_17)
-        }
-    }
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach { iosTarget ->
-        iosTarget.binaries.framework {
-            baseName = "ComposeApp"
-            isStatic = true
-        }
-    }
-
-    jvm("desktop")
-
     sourceSets {
         val desktopMain by getting
 
@@ -61,66 +24,9 @@ kotlin {
             implementation(libs.koin.core)
             implementation(libs.bundles.koin.compose)
         }
+
         desktopMain.dependencies {
-            
-        }
-    }
-}
 
-android {
-    namespace = "com.kotlearn.minesweeperk"
-    compileSdk = libs.versions.android.compileSdk.get().toInt()
-
-    defaultConfig {
-        applicationId = "com.kotlearn.minesweeperk"
-        minSdk = libs.versions.android.minSdk.get().toInt()
-        targetSdk = libs.versions.android.targetSdk.get().toInt()
-        versionCode = versionCodeProperty
-        versionName = versionNameProperty
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    buildTypes {
-        getByName("release") {
-            isMinifyEnabled = false
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-}
-
-dependencies {
-    debugImplementation(compose.uiTooling)
-}
-
-compose.desktop {
-    application {
-        mainClass = "com.kotlearn.minesweeperk.MainKt"
-
-        nativeDistributions {
-            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
-            packageName = "MinesweeperK"
-            packageVersion = versionNameProperty
-            description = "MinesweeperK by Kotlearn"
-            copyright = "Copyright (c) 2025, Kotlearn"
-            licenseFile.set(project.file("../LICENSE.txt"))
-
-            macOS {
-                dockName = "MinesweeperK"
-                entitlementsFile.set(project.file("default.entitlements"))
-            }
-        }
-
-        buildTypes.release {
-            proguard {
-                obfuscate.set(true)
-                configurationFiles.from("proguard-rules.pro")
-            }
         }
     }
 }
