+++ composeApp/src/commonMain/kotlin/com/kotlearn/minesweeperk/App.kt
@@ -1,6 +1,6 @@
 package com.kotlearn.minesweeperk
 
-import androidx.compose.material.MaterialTheme
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
