+++ feature/highscores/build.gradle.kts
@@ -1,37 +1,12 @@
-import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-
 plugins {
-    alias(libs.plugins.kotlinMultiplatform)
-    alias(libs.plugins.androidLibrary)
-    alias(libs.plugins.composeMultiplatform)
-    alias(libs.plugins.composeCompiler)
+    alias(libs.plugins.kotlearn.kotlinMultiplatform)
+    alias(libs.plugins.kotlearn.androidLibrary)
+    alias(libs.plugins.kotlearn.composeMultiplatform)
     alias(libs.plugins.kotlinSerialization)
 }
 
 kotlin {
-    androidTarget {
-        @OptIn(ExperimentalKotlinGradlePluginApi::class)
-        compilerOptions {
-            jvmTarget.set(JvmTarget.JVM_17)
-        }
-    }
-
-    listOf(
-        iosX64(),
-        iosArm64(),
-        iosSimulatorArm64()
-    ).forEach { iosTarget ->
-        iosTarget.binaries.framework {
-            baseName = "Highscores"
-            isStatic = true
-        }
-    }
-
-    jvm("desktop")
-
     sourceSets {
-
         commonMain.dependencies {
             implementation(projects.ui.core)
             implementation(projects.domain.game)
@@ -41,28 +16,5 @@ kotlin {
             implementation(libs.kotlinx.serialization)
         }
 
-    }
-}
-
-android {
-    namespace = "com.kotlearn.minesweeperk.feature.highscores"
-    compileSdk = libs.versions.android.compileSdk.get().toInt()
-
-    defaultConfig {
-        minSdk = libs.versions.android.minSdk.get().toInt()
-    }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
-    buildTypes {
-        getByName("release") {
-            isMinifyEnabled = false
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
     }
 }
