+++ feature/highscores/src/commonMain/kotlin/com/kotlearn/minesweeperk/feature/highscores/HighscoresScreen.kt
@@ -1,7 +1,7 @@
 package com.kotlearn.minesweeperk.feature.highscores
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.material.Text
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
