+++ feature/menu/src/commonMain/kotlin/com/kotlearn/minesweeperk/feature/menu/MenuButton.kt
@@ -1,7 +1,7 @@
 package com.kotlearn.minesweeperk.feature.menu
 
-import androidx.compose.material.Text
-import androidx.compose.material.TextButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
