+++ feature/play/src/commonMain/kotlin/com/kotlearn/minesweeperk/feature/play/MinesweeperBoard.kt
@@ -3,10 +3,10 @@ package com.kotlearn.minesweeperk.feature.play
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Icon
-import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Flag
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
