+++ feature/settings/src/commonMain/kotlin/com/kotlearn/minesweeperk/feature/settings/SettingsScreen.kt
@@ -1,9 +1,9 @@
 package com.kotlearn.minesweeperk.feature.settings
 
 import androidx.compose.foundation.layout.Column
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.material.TextField
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
