+++ gradle.properties
@@ -8,3 +8,5 @@ org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8
 #Android
 android.nonTransitiveRClass=true
 android.useAndroidX=true
+
+kotlin.native.ignoreDisabledTargets=true
